ENH Support for 1-bit TIFFs
The interface with numpy is a bit of a hack, but it seems to actually
work well. In general, this is a bit hard to keep both (1) generic and
(2) elegant. Numpy does not support 1-bit arrays, so I always need to
expand them, but I don't want to have that operation for all possible
image types.